requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.1.3
pillow==11.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.0
urllib3==2.2.3
